Creating components for encapsulation of menu and drag and drop functionality;
